Add placement plan feature for product deliveries
Introduced a new PlacementItem model and implemented logic to calculate efficient warehouse placement routes for delivered products. Added a new /placement-plan endpoint in DeliveryController to expose this functionality, and adapted related repositories and DTOs to support the feature.

+++ src/main/java/be/vdab/scrumjava202504/products/SimpleProductDTO.java
@@ -3,9 +3,9 @@ package be.vdab.scrumjava202504.products;
 public class SimpleProductDTO {
     private String name;
     private String ean;
-    private int productId;
+    private long productId;
 
-    public SimpleProductDTO(String ean, int productId, String name) {
+    public SimpleProductDTO(String ean, long productId, String name) {
         this.name = name;
         this.ean = ean;
         this.productId = productId;
@@ -19,7 +19,7 @@ public class SimpleProductDTO {
         return ean;
     }
 
-    public int getProductId() {
+    public long getProductId() {
         return productId;
     }
 }
